requirements.txt
pypdf
fastapi
python-dotenv
google-generativeai
pdf2image
pytesseract
chromadb
faiss-cpu
langchain_google_genai
langchain
langchain-community
httpx
pytest